Cleanup for doc
